requirements.txt
asgiref==3.7.2
base==0.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap4==22.2
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
filelock==3.10.6
fonttools==4.33.3
grade==3.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.2
libretranslatepy==2.1.1
lxml==4.9.1
MarkupSafe==2.1.3
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.27.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
translate==3.6.1
typing_extensions==4.6.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.21.0
virtualenv-clone==0.5.7
whitenoise==6.4.0
wikipedia==1.4.0
wordcloud==1.8.1
